Updated outlets and auth

+++ app/routes/outlets/outletsCRUD.js
@@ -1,5 +1,7 @@
 const createModelMiddleware = require('../middleware/createModelMiddleware');
 const findModelByMiddleware = require('../middleware/findModelByMiddleware');
+const authMiddleware = require('../../auth/middleware/authMiddleware');
+const checkEndpointAccessMiddleware = require('../../auth/middleware/checkEndpointAccessMiddleware');
 const {OutletModel} = require('../../models/appModels');
 
 module.exports = (router) => {
@@ -10,18 +12,24 @@ module.exports = (router) => {
         });
 
     router.get('/outlets/:id',
+        authMiddleware(),
+        checkEndpointAccessMiddleware(4),
         findModelByMiddleware(OutletModel, {_id: 'id'}, 'params', 'outlet'),
         (req, res) => {
             res.jsonSuccess(req.outlet);
         });
 
     router.post('/outlets',
+        authMiddleware(),
+        checkEndpointAccessMiddleware(4),
         createModelMiddleware(OutletModel, 'outlet'),
         (req, res) => {
             res.jsonSuccess(req.outlet);
         });
 
     router.patch('/outlets/:id',
+        authMiddleware(),
+        checkEndpointAccessMiddleware(4),
         findModelByMiddleware(OutletModel, {_id: 'id'}, 'params', 'outlet'),
         async (req, res) => {
             try {
@@ -36,6 +44,8 @@ module.exports = (router) => {
         });
 
     router.delete('/outlets/:id',
+        authMiddleware(),
+        checkEndpointAccessMiddleware(4),
         findModelByMiddleware(OutletModel, {_id: 'id'}, 'params', 'outlet'),
         async (req, res) => {
             await req.outlet.delete();
